+++ yfundamentals/financials.py
@@ -1,7 +1,7 @@
 # Financials
 import logging
 
-from ..common import html_wrapped_json
+from common import html_wrapped_json
 
 
 class Financials (html_wrapped_json.HTMLWrappedJson):
